+++ tests/common/data_generator/__init__.py
@@ -5,13 +5,13 @@ Copyright © 2022 Pixelgen Technologies AB.
 
 from __future__ import annotations
 
-from tests.common.test_data_generator.molecules import (
+from tests.common.data_generator.molecules import (
     generate_edgelist,
     populate_cell,
 )
-from tests.common.test_data_generator.pixelfile import write_pna_pxl
-from tests.common.test_data_generator.reads import write_pna_fastq
-from tests.common.test_data_generator.topology import generate_cell_graph
+from tests.common.data_generator.pixelfile import write_pna_pxl
+from tests.common.data_generator.reads import write_pna_fastq
+from tests.common.data_generator.topology import generate_cell_graph
 
 __all__ = [
     "generate_cell_graph",
